% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPCAT.R
\name{hypotheses}
\alias{hypotheses}
\title{Create main hypotheses and intersection hypotheses.}
\usage{
hypotheses(n)
}
\arguments{
\item{n}{The number of treatments exclusive of the control group. That is, if we have, e.g., a control group and 3 treatments choose n=3.}
}
\value{
A contrast matrix reflecting all main and intersection hypotheses induced by the closure principle (CP). Please note that the leading column of treatment 0 (i.e., the control group) is not displayed because it always contains only ones.
}
\description{
Create main hypotheses and intersection hypotheses.
}
\examples{
hypotheses(3)
hypotheses(n=2)
}
