% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Description of testdata}
\format{
A data frame with 20 rows and 2 variables:
\describe{
\item{Poissondata}{Contains 20 realizations of Poisson distributed random variables.}
\item{group}{Contains the group names of the Poisson data.}
}
}
\source{
{Artificial example data.}
}
\usage{
testdata
}
\description{
The data set testdata contains two variables.
}
\examples{
data(testdata)    #Lazy loading of the test data. To view the data type testdata.
}
\keyword{datasets}
