% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata2}
\alias{testdata2}
\title{Description of testdata2}
\format{
A data frame with 16 rows and 2 variables:
\describe{
\item{Poissondata}{Contains 16 realizations of Poisson distributed random variables.}
\item{group}{Contains the group names of the Poisson data.}
}
}
\source{
{Artificial example data.}
}
\usage{
testdata2
}
\description{
The data set testdata2 contains two variables.
}
\examples{
data(testdata2)    #Lazy loading of the test data. To view the data type testdata.
}
\keyword{datasets}
